% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_letter_to_score.R
\name{bind_letter_to_score}
\alias{bind_letter_to_score}
\title{Helper to Extract A Letter Grade From Scores}
\usage{
bind_letter_to_score(df, columns, rows)
}
\arguments{
\item{df}{A dataframe/table/spreadsheet}

\item{columns}{Range. The column indices to be targeted.}

\item{rows}{Range. The row indicies to be targeted.}
}
\value{
A dataframe with letters attached to scores
}
\description{
Helper to Extract A Letter Grade From Scores
}
\examples{
\dontrun{ #dont run because function is not exported
df <- data.frame(
 "Areas" = c("A", "B", "C", "D", "E", "F"),
 "Score1" = c(0, 20, 40, 60, 80, 100),
 "Score2" = c(0, 20, 40, 60, 80, 100),
 "Other" = c("Z", "Y", "X", "W", "V", "U")
)

bind_letter_to_score(
 df = df, 
 columns = 2:3,
 rows = 2:3
)
}

}
