% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_basins.R
\name{build_basins}
\alias{build_basins}
\title{Build Basins in the Northern Three region}
\usage{
build_basins()
}
\value{
An sf object
}
\description{
Build Basins in the Northern Three region
}
\examples{
basin_output <- build_basins()

}
