% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_land_sp.R
\name{build_land_sp}
\alias{build_land_sp}
\title{Apply Special Transformations to the n3_land Dataset}
\usage{
build_land_sp(n3_land)
}
\arguments{
\item{n3_land}{An sf object produced within the \code{\link[=build_n3_region]{build_n3_region()}} function}
}
\value{
An sf object
}
\description{
Apply Special Transformations to the n3_land Dataset
}
\examples{
n3_land <- build_land_sp(n3_land)

}
