% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_marine_sp.R
\name{build_marine_sp}
\alias{build_marine_sp}
\title{Apply Special Transformations to the n3_marine Dataset}
\usage{
build_marine_sp(n3_marine, n3_land)
}
\arguments{
\item{n3_marine}{An sf object produced within the \code{\link[=build_waterbody_boundaries]{build_waterbody_boundaries()}} function}

\item{n3_land}{An sf object produced within the \code{\link[=build_n3_region]{build_n3_region()}} function}
}
\value{
An sf object
}
\description{
Apply Special Transformations to the n3_marine Dataset
}
\examples{
\dontrun{ #dont run because function is not exported
n3_marine <- build_marine_sp(n3_marine, n3_land)
}

}
