% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_n3_region.R
\name{build_n3_region}
\alias{build_n3_region}
\title{Build the Northern Three Region Dataset}
\usage{
build_n3_region()
}
\value{
An sf object
}
\description{
Build the Northern Three Region Dataset
}
\examples{
#' n3_region <- build_n3_region()

}
