% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_named_islands.R
\name{build_named_islands}
\alias{build_named_islands}
\title{Build the Named Islands Within the Northern Three Region}
\usage{
build_named_islands(basins)
}
\arguments{
\item{basins}{An sf object produced by the \code{\link[=build_basins]{build_basins()}} function}
}
\value{
An sf object
}
\description{
Build the Named Islands Within the Northern Three Region
}
\examples{
\dontrun{ #dont run because function is not exported
named_islands <- build_named_islands(basins)
}

}
