% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sub_basins.R
\name{build_sub_basins}
\alias{build_sub_basins}
\title{Build the Sub Basins Within the Northern Three Region}
\usage{
build_sub_basins(n3_land, basins, n3_marine)
}
\arguments{
\item{n3_land}{An sf object produced within the \code{\link[=build_n3_region]{build_n3_region()}} function}

\item{basins}{An sf object produced by the \code{\link[=build_basins]{build_basins()}} function}

\item{n3_marine}{An sf object produced within the \code{\link[=build_waterbody_boundaries]{build_waterbody_boundaries()}} function}
}
\value{
An sf object
}
\description{
Build the Sub Basins Within the Northern Three Region
}
\examples{
n3_land <- build_sub_basins(n3_land, basins, n3_marine)

}
