% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_water_types.R
\name{build_water_types}
\alias{build_water_types}
\title{Build the Landbased Water Types Within the Northern Three Region}
\usage{
build_water_types(n3_land, n3_marine)
}
\arguments{
\item{n3_land}{An sf object produced within the \code{\link[=build_n3_region]{build_n3_region()}} function}

\item{n3_marine}{An sf object produced within the \code{\link[=build_waterbody_boundaries]{build_waterbody_boundaries()}} function}
}
\value{
An sf object
}
\description{
Build the Landbased Water Types Within the Northern Three Region
}
\examples{
\dontrun{ #dont run because function is not exported
n3_land <- build_water_types(n3_land, n3_marine)
}

}
