% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_waterbodies.R
\name{build_waterbodies}
\alias{build_waterbodies}
\title{Build the Marine Based Water Types Within the Northern Three Region}
\usage{
build_waterbodies(basins, named_islands)
}
\arguments{
\item{basins}{An sf object produced by the \code{\link[=build_basins]{build_basins()}} function}

\item{named_islands}{An sf object produced by the \code{\link[=build_named_islands]{build_named_islands()}} function}
}
\value{
An sf object
}
\description{
Build the Marine Based Water Types Within the Northern Three Region
}
\examples{
\dontrun{ #dont run because function is not exported
waterbodies <- build_waterbodies(basins, named_islands)
}

}
