% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_waterbody_boundaries.R
\name{build_waterbody_boundaries}
\alias{build_waterbody_boundaries}
\title{Build the Marine Zone Boundaries Within the Northern Three Region}
\usage{
build_waterbody_boundaries(waterbodies)
}
\arguments{
\item{waterbodies}{An sf object produced by the \code{\link[=build_waterbodies]{build_waterbodies()}} function}
}
\value{
An sf object
}
\description{
Build the Marine Zone Boundaries Within the Northern Three Region
}
\examples{
n3_marine <- build_waterbody_boundaries(waterbodies)

}
