% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox_suffixes.R
\name{checkbox_suffixes}
\alias{checkbox_suffixes}
\title{Checkbox Suffixes}
\usage{
checkbox_suffixes(field_names, data_dict)
}
\arguments{
\item{field_names}{The current field names of interest}

\item{data_dict}{The meta data data frame.}
}
\description{
Checkbox variables return one vector of data for each option defined
in the variable.  The variables are returned with the suffix \code{___[option]}.
\code{rc_export} needs these suffixes in order to retrieve all of the
variables and to apply the correct labels.
}
