% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_form_rc_grades.R
\name{cond_form_rc_grades}
\alias{cond_form_rc_grades}
\title{Title}
\usage{
cond_form_rc_grades(df, file_name, target_columns, target_rows, include_letter)
}
\arguments{
\item{df}{A standard dataframe, data.frame, table, etc.}

\item{file_name}{String. The name to save the output under, without an extension}

\item{target_columns}{Numeric Vector. The columns colour grading should be applied to. e.g. 1:10}

\item{target_rows}{Numeric Vector. The rows colour grading should be applied to. e.g. 1:10}

\item{include_letter}{Boolean. Should a letter grade be attached to the score.}
}
\value{
A .xlsx output
}
\description{
Title
}
\examples{
x <- data.frame(
 Basin = c("Black", "Ross", "Haughton", "Suttor"),
 DIN = c(51, 76, 27, 98),
 TP = c(90, 57, 34, 72)
)
cond_form_rc_grades(df = x, file_name = "final_scores", target_columns = 2:3, target_rows = 1:4, include_letter = FALSE)

}
