% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_formatter.R
\name{conditional_formatter}
\alias{conditional_formatter}
\title{A Helper That Presents Conditional Formatting Options}
\usage{
conditional_formatter(
  data,
  workbook,
  target_columns,
  target_rows,
  scheme,
  include_letter,
  file_name
)
}
\arguments{
\item{data}{The main dataframe to run functions on}

\item{workbook}{An excel workbook object.}

\item{target_columns}{Numeric Vector. The columns that colour grading should be applied to. e.g. 1:10.}

\item{target_rows}{Numeric Vector. The rows that colour grading should be applied to. e.g. 1:10.}

\item{scheme}{String. Defines the conditional formatting rules and colours. Options are Report Card Grade, Report Card Score, Rainfall Temperature, Summary Stat}

\item{include_letter}{Boolean. Should a letter grade be attached to the score? This is specific to Report Card scores}

\item{file_name}{String. The name of the file you want to save. Without the extension}
}
\value{
An excel workbook object
}
\description{
A Helper That Presents Conditional Formatting Options
}
\examples{
\dontrun{
conditional_formatter(
data, workbook, target_columns, 
target_rows, scheme, include_letter, file_name)
}
}
