% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_to_string.R
\name{data_frame_to_string}
\alias{data_frame_to_string}
\title{Converts dataframe to csv string}
\usage{
data_frame_to_string(data)
}
\arguments{
\item{data}{Dataframe}
}
\description{
Converts dataframe to csv string
}
