% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_colour_scheme.R
\name{define_colour_scheme}
\alias{define_colour_scheme}
\title{Helper to Select Various Colour Schemes}
\usage{
define_colour_scheme(workbook, scheme)
}
\arguments{
\item{workbook}{A excel workbook object created by openxlsx2}

\item{scheme}{String. One of several options: Report Card, Rainfall, Temperature, Summary Stat, TBD.}
}
\value{
The supplied workbook object, now with the colour scheme embedded
}
\description{
Helper to Select Various Colour Schemes
}
\examples{
\dontrun{ #dont run because function is not exported
wb <- openxlsx2::wb_workbook()
define_colour_scheme(workbook = wb, scheme = "Report Card")
}

}
