% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_polygon_highlight.R
\name{dem_polygon_highlight}
\alias{dem_polygon_highlight}
\title{Highlight a specific area defined by a polygon}
\usage{
dem_polygon_highlight(Highlight, Extent, MapArray, MapMatrix)
}
\arguments{
\item{Highlight}{An sf object that defines the area of interest. Must be within map boundaries}

\item{Extent}{An sf object that defines the full map boundaries}

\item{MapArray}{An array object created specifically by the 'dem_base_map()' function}

\item{MapMatrix}{A matrix object created specifically by the 'dem_base_map()' function}
}
\value{
The provided array object with highlights now embedded.
}
\description{
Highlight a specific area defined by a polygon
}
\examples{
\dontrun{ #dont run because function takes a long time

maggie <- build_n3_region() |> 
    filter(SubBasinOrSubZone == "Magnetic Island")

map_ext <- ext(my_spat_raster)

map_array <- dem_base_map(...)

dem_cropped <- dem_polygon_highlight(
  Highlight = maggie,
  Extent = map_ext,
  MapArray = map_array
)
}

}
