% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ereef_aggregation_helper.R
\name{ereefs_aggregation_helper}
\alias{ereefs_aggregation_helper}
\title{A function to help with various forms of netCDF aggregation}
\usage{
ereefs_aggregation_helper(nc, agg)
}
\arguments{
\item{nc}{A netCDF object}

\item{agg}{The type of aggregation to complete, defaults to "Month". Options are "Month", "Season", "Financial", "Annual"}
}
\value{
A netCDF object
}
\description{
A function to help with various forms of netCDF aggregation
}
\examples{
\dontrun{
x <- ereefs_aggregation_helper(nc, "Month")
}
}
