% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ereefs_dotplot.R
\name{ereefs_dotplot}
\alias{ereefs_dotplot}
\title{Plotting eReefs Data}
\usage{
ereefs_dotplot(
  nc,
  SubSample = 500,
  Heading = NULL,
  YAxisName = NULL,
  LogTransform = FALSE
)
}
\arguments{
\item{nc}{A NetCDF (stars) object or a list of NetCDF objects, generally produced by the \code{\link[=ereefs_extract]{ereefs_extract()}} function}

\item{SubSample}{Numeric String. The number of values per day to plot in the dot plot. Defaults to 500 values per day.}

\item{Heading}{Character String. The heading of the plot. Defaults to NULL.}

\item{YAxisName}{Character String. The yaxis label text for the plot. Defaults to NULL and returns "value" in the plot.}

\item{LogTransform}{Boolean. Should the data be presented with a log transformation - useful for some variables such as Turbidity or Chlorophyll}
}
\value{
A ggplot object
}
\description{
Plotting eReefs Data
}
\examples{

nc <- system.file("extdata/turbidity22.nc", package = "RcTools")
nc <- stars::read_mdim(nc)

x_plot <- ereefs_dotplot(
  nc = nc,
  YAxisName = "Turbidity (NTU)", 
  Heading = "Dry Tropics", 
  LogTransform = TRUE
)

}
