% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ereefs_extract.R
\name{ereefs_extract}
\alias{ereefs_extract}
\title{Extracting eReefs Data}
\usage{
ereefs_extract(
  Model = "catalog",
  Region,
  StartDate,
  EndDate,
  Variable,
  Downsample = 0
)
}
\arguments{
\item{Model}{Character String. Describes the url of the model to use. Defaults to "catalog" and provides the user with options}

\item{Region}{SF Object. An sf object defining the area to extract data from}

\item{StartDate}{POSIX time. Supplied as a string in the format YYYY-MM-DD that defines the first day to extract data from}

\item{EndDate}{POSIX time. Supplied as a string in the format YYYY-MM-DD that defines the last day to extract data from}

\item{Variable}{Character Vector. The variable to extract from eReefs, currently accessible: "Turbidity", "Chlorophyll a", "DIN", "NH4", "NO3", "Secchi", "pH", "Wind", "True Colour"}

\item{Downsample}{Integer. A singular value that defines the level of downsampling to apply to the data (reduces data size and processing time, also reduces resolution)}
}
\value{
A single NetCDF (stars) object of specified dimensions and attributes
}
\description{
Extracting eReefs Data
}
\examples{
\dontrun{ #dont run because function takes a while to execute and calls interactive component
sf_obj <- system.file("extdata/boundary.gpkg", package = "RcTools")
sf_obj <- sf::st_read(sf_obj)

nc <- ereefs_extract(
  Region = sf_obj,
  StartDate = "2022-03-01",
  EndDate = "2022-03-03",
  Variable = "Turbidity",
  Downsample = 0
)
}

}
