% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ereefs_get_date_range.R
\name{ereefs_get_date_range}
\alias{ereefs_get_date_range}
\title{Return the Model's Min and Max Dates}
\usage{
ereefs_get_date_range(Model = "catalog")
}
\arguments{
\item{Model}{Character string. Defaults to "catalog" however can also accept a premade url}
}
\value{
Character string. Describes the min and max dates for the specific model
}
\description{
Return the Model's Min and Max Dates
}
\examples{
date_range <- ereefs_get_date_range()

}
