% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ereefs_get_palette.R
\name{ereefs_get_palette}
\alias{ereefs_get_palette}
\title{Return Colour Palettes Associated With Specific Variables}
\usage{
ereefs_get_palette(Variable)
}
\arguments{
\item{Variable}{Character String. The Variable being mapped, currently accessible: "Turbidity", "Chlorophyll a", "DIN", "NH4", "NO3", "Secchi", "pH", "Wind", "True Colour"}
}
\value{
A vector of hex code colour values of length 9
}
\description{
Return Colour Palettes Associated With Specific Variables
}
\examples{
\dontrun{
ereefs_get_palette("Turbidity")
}
}
