% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ereefs_input_selection.R
\name{ereefs_input_selection}
\alias{ereefs_input_selection}
\title{Provide a Range of Models to Choose From}
\usage{
ereefs_input_selection(input_file = "catalog")
}
\arguments{
\item{input_file}{Character string. Defaults to "catalog" however can also accept a premade url}
}
\value{
A Character string url to one of 12 models.
}
\description{
Provide a Range of Models to Choose From
}
\examples{
\dontrun{ #dont run because function is not exported
url <- ereefs_input_selection()
}

}
