% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ereefs_list_safety_check_and_convert.R
\name{ereefs_list_safety_check_and_convert}
\alias{ereefs_list_safety_check_and_convert}
\title{Check if object is Stars or list of Stars, Convert if List}
\usage{
ereefs_list_safety_check_and_convert(nc)
}
\arguments{
\item{nc}{A list of NetCDF (stars) objects, generally produced by the \code{\link[=ereefs_extract]{ereefs_extract()}} function}
}
\value{
A single NetCDF (stars) object combined on the time dimension
}
\description{
Check if object is Stars or list of Stars, Convert if List
}
\examples{
\dontrun{ #dont run because function is not exported
ereefs_list_safety_check_and_convert("Turbidity")
}

}
