% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ereefs_map.R
\name{ereefs_map}
\alias{ereefs_map}
\title{Mapping eReefs Data}
\usage{
ereefs_map(nc, MapType, Aggregation, LegendTitle = NULL, nrow = NULL)
}
\arguments{
\item{nc}{A netCDF (stars) object, usually produced by the \link{ereefs_extract} and \link{ereefs_reproject} functions}

\item{MapType}{Character String. Defines the type of map produced. One of "Concentration", "True Colour", or "Vector Field" (AKA wind)}

\item{Aggregation}{Character String. Defines the level of aggregation to apply. Defaults to "Month". Options are "Month", "Season", "Financial", "Annual"}

\item{LegendTitle}{Character String. The title of the legend. Defaults to the name of the netCDF's attribute}

\item{nrow}{Numeric String. The number of rows to be used when facetting maps. Defaults to NULL and uses underlying default value}
}
\value{
A tmap object
}
\description{
Mapping eReefs Data
}
\examples{

nc <- system.file("extdata/turbidity22.nc", package = "RcTools")
nc <- stars::read_mdim(nc)

m <- ereefs_map(
  nc = nc,
  MapType = "Concentration",
  Aggregation = "Month"
)

}
