% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ereefs_plot.R
\name{ereefs_plot}
\alias{ereefs_plot}
\title{Convert the eReefs object from curvilinear to a regular grid}
\usage{
ereefs_plot(
  nc,
  SubSample = 500,
  Heading = NULL,
  YAxisName = NULL,
  LogTransform = FALSE
)
}
\arguments{
\item{nc}{A NetCDF (stars) object or a list of NetCDF objects, generally produced by the \code{\link[=ereefs_extract]{ereefs_extract()}} function}

\item{SubSample}{Numeric String. The number of values per day to plot in the dot plot. Defaults to 500 values per day.}

\item{LogTransform}{Boolean. Should the data be presented with a log transformation - useful for some variables such as Turbidity or Chlorophyll}

\item{Variable}{Character String. The variable in the netCDF that is being mapped. One of: "Turbidity", "Chlorophyll a", "DIN", "NH4", "NO3", "Secchi", "pH", "Wind".}

\item{Region}{Character String. The area in which the netCDF is located.}
}
\value{
A ggplot object
}
\description{
Convert the eReefs object from curvilinear to a regular grid
}
\examples{
\dontrun{
x_plot <- ereefs_plot(nc, YAxisName = "Turbidity (NTU)", Heading = "Dry Tropics", LogTransform = TRUE)
}
}
