% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ereefs_reproject.R
\name{ereefs_reproject}
\alias{ereefs_reproject}
\title{Convert the eReefs object from curvilinear to a regular grid}
\usage{
ereefs_reproject(x)
}
\arguments{
\item{x}{A NetCDF (stars) object, generally produced by the \code{\link[=ereefs_extract]{ereefs_extract()}} function}
}
\value{
A NetCDF (stars) object with the same values as x, but in a regular grid format
}
\description{
Convert the eReefs object from curvilinear to a regular grid
}
\examples{
\dontrun{
x_regular <- ereefs_reproject(x_curvilinear)
}
}
