% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ereefs_windrose.R
\name{ereefs_windrose}
\alias{ereefs_windrose}
\title{Create a Windrose Plot from eReefs Data}
\usage{
ereefs_windrose(
  nc,
  SubSample = 500,
  Aggregation = "Month",
  Heading = "Approximated Wind Speed",
  LegendTitle = "Speed (km/h)"
)
}
\arguments{
\item{nc}{A single NetCDF (stars) object OR a list of NetCDF (stars) objects, generally produced by the \code{\link[=ereefs_extract]{ereefs_extract()}} function}

\item{SubSample}{Numeric String. The number of values per day to plot in the wind rose. Defaults to 500 values per day.}

\item{Aggregation}{Character String. What type of grouping to apply to the data. Defaults to "Month". Options are "Month", "Season", "Financial", "Annual"}

\item{Heading}{Character String. The heading of the plot. Defaults to "Approximated Wind Speed".}

\item{LegendTitle}{Character String. The title of the legend. Defaults to "Speed (km/h)".}
}
\value{
A ggplot2 object (a windrose plot)
}
\description{
Create a Windrose Plot from eReefs Data
}
\examples{

nc_names <- paste0("extdata/", c("wind_dir", "wind_mag", "wind_u", "wind_v"), ".nc")
nc <- purrr::map(nc_names, \(x) system.file(x, package = "RcTools"))
nc <- purrr::map(nc, stars::read_mdim)

wr_plot <- ereefs_windrose(nc)

}
