% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportProjectInformation.R
\name{exportProjectInformation}
\alias{exportProjectInformation}
\alias{exportProjectInformation.redcapApiConnection}
\alias{exportProjectInformation.redcapDbConnection}
\title{Exports the Project Information}
\usage{
exportProjectInformation(
  url = getOption("redcap_bundle")$redcap_url,
  token = getOption("redcap_token"),
  ...,
  error_handling = getOption("redcap_error_handling")
)
}
\arguments{
\item{url}{A url address to connect to the REDCap API}

\item{token}{A REDCap API token}

\item{...}{Arguments to be passed to other methods.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcap_error}}}
}
\description{
Retrieve a data frame with the project information.
}
\details{
If this function is used in a version of REDCap that does not
support the Export Version Number function, the character string
\code{'Version Unknown'} is returned.

REDCap API Documentation (6.5.0):
This function allows you to export some of the basic attributes of a given
REDCap project, such as the project's title, if it is longitudinal, if
surveys are enabled, the time the project was created and moved to production, etc.

REDCap Version:
6.5.0 (Perhaps earlier)

Known REDCap Limitations:
None
}
\references{
Please refer to your institution's API documentation.

Additional details on API parameters are found on the package wiki at
\url{https://github.com/nutterb/redcapAPI/wiki/REDCap-API-Parameters}
}
\author{
Benjamin Nutter
}
