% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_air_temp.R
\name{extract_air_temp}
\alias{extract_air_temp}
\title{Extract, Crop, and Combine Air Temperature Data}
\usage{
extract_air_temp(MinPath, MaxPath, CropObj)
}
\arguments{
\item{MinPath}{A character string. Defines the full path to the folder in which all min temp sub folders are stored.
Generally this will be user/.../.../raw/tmin2/}

\item{MaxPath}{A character string. Defines the full path to the folder in which all max temp sub folders are stored.
Generally this will be user/.../.../raw/tmax2/}

\item{CropObj}{Sf Object. An sf object used to define the area in which data is to be cropped to. Generally,
the n3_region object from the \code{\link[=build_n3_region]{build_n3_region()}} function is used.}
}
\value{
A list of three netCDF objects. Specifically a min temp, max temp, and mean temp netCDF
}
\description{
Extract, Crop, and Combine Air Temperature Data
}
\examples{
\dontrun{ #dont run because function takes a while

min_path <- "C:/Users/path/to/data/raw/tmin2"
max_path <- "C:/Users/path/to/data/raw/tmax2"

n3_region <- build_n3_region()

nc <- extract_air_temp(n3_region, min_path, max_path)

}

}
