% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_land_use.R
\name{extract_land_use}
\alias{extract_land_use}
\title{Extract Land Use Data from a collection of Datasets on File}
\usage{
extract_land_use(RawPath, CroppedPath, CropObj)
}
\arguments{
\item{RawPath}{Character String. A path arguement to the location where each full file should be retrived from.}

\item{CroppedPath}{Character String. A path arguement to the location where each cropped file should be saved and retrived.}

\item{CropObj}{Sf Object. An sf object used to define the area in which data is to be cropped to. Generally,
the n3_region object from the \code{\link[=build_n3_region]{build_n3_region()}} function is used.}
}
\value{
A single sf object.
}
\description{
Extract Land Use Data from a collection of Datasets on File
}
\examples{
\dontrun{ #dont run because function takes a long time to process

p1 <- "path/to/folder/raw/"
p2 <- "path/to/folder/processed/"

n3_region <- build_n3_region()

lu_data <- extract_land_use(p1, p2, n3_region)
}

}
