% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rainfall.R
\name{extract_rainfall}
\alias{extract_rainfall}
\title{Extract Rainfall Data from BOM AWO API}
\usage{
extract_rainfall(CropObj, StartDate = NULL, EndDate = NULL)
}
\arguments{
\item{CropObj}{Sf Object. An sf object used to define the area in which data is to be cropped to. Generally,
the n3_region object from the \code{\link[=build_n3_region]{build_n3_region()}} function is used.}

\item{StartDate}{A character string in the format of YYYY-MM-DD. Defaults to "1911-01-31" (earliest value)}

\item{EndDate}{A character string in the format of YYYY-MM-DD. Defaults to newest value}
}
\value{
A stars  netCDF object
}
\description{
Extract Rainfall Data from BOM AWO API
}
\examples{
\dontrun{ #dont run because example takes a while
n3_region <- build_n3_region()

rain_data <- extract_rainfall(n3_region, "2022-01-01", "2022-02-01")
}

}
