% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sst.R
\name{extract_sst}
\alias{extract_sst}
\title{Extract Sea Surface Temperature Data from NOAA's API}
\usage{
extract_sst(RawPath, CroppedPath, CropObj)
}
\arguments{
\item{RawPath}{Character String. A path arguement to the location where each full file should be saved and retrived.}

\item{CroppedPath}{Character String. A path arguement to the location where each cropped file should be saved and retrived.}

\item{CropObj}{Sf Object. An sf object used to define the area in which data is to be cropped to. Generally,
the n3_region object from the \code{\link[=build_n3_region]{build_n3_region()}} function is used.}
}
\value{
A stars netCDF object AND saves individual laers to file. (Note, this is because each file can only be downloaded
individually. If just the single netCDF was returned, every layer would need to be re-downloaded once a new layer is
published. Instead, only the new layer is downloaded, and the single netCDF is rebuilt locally).
}
\description{
Extract Sea Surface Temperature Data from NOAA's API
}
\examples{
\dontrun{ #dont run because function takes a long time to process

p1 <- "path/to/folder/raw/"
p2 <- "path/to/folder/processed/"

n3_region <- build_n3_region()

sst_data <- extract_sst(p1, p2, n3_region)
}

}
