% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_functions.R
\name{format_variables}
\alias{format_variables}
\title{Convert a REDCap Data Field to an R Vector}
\usage{
format_variables(record_data, data_dict, factor_labels = TRUE, dates = TRUE)
}
\arguments{
\item{record_data}{A data frame of records returned by \code{rc_export}}

\item{data_dict}{A data frame giving the data data_dictionary, as returned
by \code{exportMetaData}}

\item{factor_labels}{Logical, determines if checkbox, radio button, dropdown and yesno
variables are labelled.}

\item{dates}{Logical, determines if date variables are converted to POSIXct format}
}
\description{
Converts a field exported from REDCap into a valid R vector
}
\author{
Jeffrey Horner
}
