% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quarto_params.R
\name{get_quarto_params}
\alias{get_quarto_params}
\title{Return the YAML params during an interative session}
\usage{
get_quarto_params()
}
\value{
a list of all YAML parameters
}
\description{
Return the YAML params during an interative session
}
\examples{

params <- get_quarto_params()

}
