% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getID.R
\name{getID}
\alias{getID}
\title{Retrieve record_id field name}
\usage{
getID(
  record_data = NULL,
  data_dict = getOption("redcap_bundle")$data_dict,
  id_field = getOption("redcap_bundle")$id_field
)
}
\arguments{
\item{record_data}{Dataframe. Record data exported from REDCap}

\item{data_dict}{Dataframe. REDCap project data data_dictionary}

\item{id_field}{Character. Field name corresponding to the 'record_id' field.}
}
\description{
Gets record_id field name from bundle, or data_dict
if provided. Otherwise, the first column name will be used. This
is an internal rctools function only.
}
\author{
Marcus Lehr
}
