% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_water_layer.R
\name{maps_water_layer}
\alias{maps_water_layer}
\title{Create a Tmap Water Layer}
\usage{
maps_water_layer(
  Basin,
  WaterLines = TRUE,
  WaterAreas = FALSE,
  WaterLakes = FALSE,
  StreamOrder = NULL
)
}
\arguments{
\item{Basin}{Character Vector. Defines the area in which the water layer should be created.}

\item{WaterLines}{Boolean. Should water lines be included in the map layer? Defaults to TRUE.}

\item{WaterAreas}{Boolean. Should water areas be included in the map layer? Defaults to TRUE.}

\item{WaterLakes}{Boolean. Should lakes be included in the map layer? Defaults to TRUE.}

\item{StreamOrder}{Numeric Vector of length 1 or 2. Defines the minimum, or minimum and maximum, stream order to filter the WaterLines
paramter by. Where 1 = the smallest streams and increasing numbers correspond to increasing stream sizes. Defaults to NULL (all).}
}
\value{
A Tmap Object
}
\description{
Create a Tmap Water Layer
}
\examples{

ross_and_black <- maps_water_layer(Basin = c("Black", "Ross"))

}
