% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingSummary.R
\name{missingSummary}
\alias{missingSummary}
\alias{missingSummary.redcapApiConnection}
\alias{missingSummary.redcapDbConneciton}
\alias{missingSummary_offline}
\title{Report of Missing Values}
\usage{
missingSummary(
  url = getOption("redcap_bundle")$redcap_url,
  token = getOption("redcap_token"),
  excludeMissingForms = TRUE,
  ...,
  proj = NULL,
  batch.size = -1
)

missingSummary_offline(records, data_dict, excludeMissingForms = TRUE)
}
\arguments{
\item{url}{A url address to connect to the REDCap API}

\item{token}{A REDCap API token}

\item{excludeMissingForms}{If all of the fields in a form are missing, would
you like to assume that they are purposefully missing?  For instance, if
a patient did not experience an adverse event, the adverse event form would
contain no data and you would not want it in this report.}

\item{...}{Additional arguments to pass to other methods.  Currently ignored.}

\item{proj}{A redcapProjectInfo object.}

\item{batch.size}{Batch size parameter for \code{rc_export}}

\item{records}{a filename pointing to the raw records download from REDCap}

\item{data_dict}{a filename pointing to the data data_dictionary download from REDCap}
}
\description{
Returns a data frame of subject events with missing values.
}
\details{
The intention of this function is to generate a list of subject
events that are missing and could potentially be values that should have
been entered.

The branching logic from the data data_dictionary is parsed and translated into
and R expression.  When a field with branching logic passes the logical
statement, it is evaluated with \code{is.na}, otherwise, it is set to
\code{FALSE} (non-missing, because there was never an opportunity to
provide a value).

Optionally, forms that are entirely missing can be determined to be
non-missing.  This is applicable when, for instance, a patient did not
have an adverse event.  In this case, a form dedicated to adverse events
would contain meaningless missing values and could be excluded from the
report.
}
\author{
Benjamin Nutter
}
