% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_high_res_crop.R
\name{nc_high_res_crop}
\alias{nc_high_res_crop}
\title{Crop A Raster and Increase its Resolution at the Same Time}
\usage{
nc_high_res_crop(nc, CropObj, DisaggFactor = 5)
}
\arguments{
\item{nc}{A single NetCDF (stars) object, usually produced as part of the Climate workflow.}

\item{CropObj}{An sf object that defines the area in which data is to be cropped too.}

\item{DisaggFactor}{Numeric string. Defines the level of disaggregation to apply to the data (i.e. how many additional cells to create).
Defaults to five, i.e. one cell will become five cells.}
}
\value{
A singele NetCDF (stars) object.
}
\description{
Crop A Raster and Increase its Resolution at the Same Time
}
\examples{

nc <- stars::read_mdim(system.file("extdata/turb_reg.nc", package = "RcTools"))
nc <- ereefs_reproject(nc)
area <- sf::st_read(system.file("extdata/example_box.gpkg", package = "RcTools"))
area <- sf::st_transform(area, sf::st_crs(nc))

nc_high_res <- nc_high_res_crop(nc, area, 5)

}
