% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_only.R
\name{numeric_only}
\alias{numeric_only}
\title{Filters REDCap records data for numeric data only}
\usage{
numeric_only(
  record_data,
  data_dict = getOption("redcap_bundle")$data_dict,
  grouping_variable = NULL,
  fields = NULL,
  long_format = TRUE,
  drop_message = TRUE
)
}
\arguments{
\item{record_data}{Dataframe. Records data export from REDCap. For the
purposes of this function, only quantitative data will be kept.}

\item{data_dict}{Dataframe. REDCap project data data_dictionary. By default,
$data_dict is expected in the REDCap bundle option, as created by
\code{rc_bundle}.}

\item{grouping_variable}{String. Name of a variable to group records by. If
included, it will be used as one of the melting factors.}

\item{fields}{Character. A vector of field/variable names to be analyzed
may be passed manually.}

\item{long_format}{Logical. Determines whether the returned dataframe will
be in long or wide format. Default is \code{TRUE}.}

\item{drop_message}{Logical. Determine if a message is shown to the user about
dropping non-numerical data}
}
\description{
Unless a vector of variables/field names is passed to the
\code{fields} argument, the fields to be analyzed will be guessed based on
column type. Furthermore, all non-numeric data will be removed before analysis.
If mixed numeric/non-numeric data (e.g. "160 cm") are passed, the first numerical
instance will be extracted from the data. A long format dataframe is returned.

This is an internal function only.
}
\author{
Marcus Lehr
}
