% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.R
\name{plot_missing}
\alias{plot_missing}
\title{Plot data from rc_missing}
\usage{
plot_missing(missing_data)
}
\arguments{
\item{missing_data}{Dataframe. A frame of long format missing data, as
created by \code{rc_missing}.}
}
\description{
Creates a pie chart for each variable passed from
\code{rc_missing}. These plots are useful for investigating the
structure of missing data. This function can be called from the
\code{plot} argument in \code{rc_missing}. Also try the \code{table}
argument in \code{rc_missing} for further investigation.

This is an internal function only.
}
\author{
Marcus Lehr
}
