% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qld.R
\docType{data}
\name{qld}
\alias{qld}
\title{Queensland Outline}
\format{
An sf object.
}
\source{
This outline was originally sourced from the gisaimsr package using:
get(data("gbr_feat", package = "gisaimsr")) |> filter(FeatName \%in\% c("Mainland", "Island"))
}
\usage{
qld
}
\description{
A spatial object containing the outline of Queensland.
}
\keyword{datasets}
