% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall_extract.R
\name{rainfall_extract}
\alias{rainfall_extract}
\title{Extract Rainfall Data from BOM AWO API}
\usage{
rainfall_extract(Region, StartDate = NULL, EndDate = NULL)
}
\arguments{
\item{Region}{An sf object}

\item{StartDate}{A character string in the format of YYYY-MM-DD. Defaults to "1911-01-31" (earliest value)}

\item{EndDate}{A character string in the format of YYYY-MM-DD. Defaults to newest value}
}
\value{
A stars  netCDF object
}
\description{
Extract Rainfall Data from BOM AWO API
}
\examples{
\dontrun{ #dont run because example takes a while
n3_region <- build_n3_region()

rain_data <- rainfall_extract(n3_region, "2022-01-01", "2022-02-01")
}

}
