% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_bundle.R
\name{rc_bundle}
\alias{rc_bundle}
\title{Export REDCap project metadata}
\usage{
rc_bundle(
  url,
  token,
  create_options = TRUE,
  return_object = TRUE,
  data_dict = FALSE,
  users = FALSE,
  instruments = FALSE,
  event_data = FALSE,
  arms = FALSE,
  mappings = FALSE,
  proj_info = FALSE,
  version = FALSE,
  dates = TRUE,
  labels = TRUE
)
}
\arguments{
\item{url}{A url address to connect to the REDCap API}

\item{token}{Path to a text file containing your REDCap API token}

\item{create_options}{Logical. Indicates whether the REDCap bundle, token, and
url should be uploaded to the R session options.}

\item{return_object}{Logical. Indicates whether the REDCap bundle should be
returned as an object.}

\item{data_dict}{Logical.  Indicates if the meta data (data data_dictionary)
should be exported.}

\item{users}{Logical. Indicates if the users table should be exported.}

\item{instruments}{Logical. Indicates if the instruments table should be exported.}

\item{event_data}{Logical. Indicates if the event names should be exported.}

\item{arms}{Logical. Indicates if the arms table should be exported.}

\item{mappings}{Logical. Indicates if the form-event mappings should
be exported.}

\item{proj_info}{Logical. Indicates if the project information should be exported.
If the project is not longitudinal, the events, arms, and event-form mappings
elements will be assigned character vectors instead of data frames.}

\item{version}{Logical. Indicates if the REDCap version number should be exported.
Only applicable in REDCap 6.0.0 and higher.}

\item{dates}{Logical. If \code{TRUE} (Default), user expiration dates are converted to
\code{POSIXct} objects.}

\item{labels}{Logical.  If \code{TRUE} (Default), the user form permissions are
converted to labelled factors.}
}
\description{
This function performs several API calls at once in order
to reduce the overall number of calls when using rctools. All project
metadata is exported and saved to a REDCap bundle option and/or object
for use with other rctools functions.
}
\details{
By default, this function exports all metadata. If any of the
metadata arguments are set to \code{TRUE}, then only the indicated data
will be exported. Additionally, by default the exported data will be
saved to the option "redcap_bundle", and other rctools functions will check
this option for metadata by default. The bundle can be accessed via
\code{getOption("redcap_bundle")}. It will also be returned as
an object (by default) for ease of access. Saving the bundle object as an
RDS (see saveRDS()) or within your workspace image (see save.image()) is
useful to negate the need for additional API calls in future sessions.

For security purposes, the REDCap token will be saved only to the option
"redcap_token". This option, along with the "redcap_bundle" option, will
not persist across R sessions. Unlike the REDCap bundle data, saving a
token within an object or workspace is not recommended.
}
\author{
Marcus Lehr

Benjamin Nutter
}
