% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_export.R
\name{rc_export}
\alias{rc_export}
\title{Export Records from a REDCap Database}
\usage{
rc_export(
  report_id = NULL,
  url = getOption("redcap_bundle")$redcap_url,
  token = getOption("redcap_token"),
  data_dict = getOption("redcap_bundle")$data_dict,
  id_field = getOption("redcap_bundle")$id_field,
  records = NULL,
  fields = NULL,
  forms = NULL,
  events = NULL,
  survey = TRUE,
  dag = TRUE,
  form_complete_auto = FALSE,
  format = FALSE,
  ...,
  strip = TRUE,
  colClasses = NA,
  batch.size = -1,
  error_handling = getOption("redcap_error_handling")
)
}
\arguments{
\item{report_id}{Numeric. ID number for a report created in REDCap.}

\item{url}{Character. A url address to connect to the REDCap API}

\item{token}{Character. A REDCap API token}

\item{data_dict}{Dataframe. REDCap project data data_dictionary. By default,
this will be fetched from the REDCap bundle option, as created by \code{rc_bundle}.
Otherwise, a data.frame containing the project data dictionary must be supplied.}

\item{id_field}{Character. The name of the record_id field for your REDCap project.}

\item{records}{Character. A vector of study id's to be returned.  If \code{NULL}, all
subjects are returned.}

\item{fields}{Character. A vector of fields to be returned.  If \code{NULL},
all fields are returned.}

\item{forms}{Character. A vector of forms to be returned.  If \code{NULL},
all forms are returned.}

\item{events}{Character. A vector of events to be returned from a longitudinal database.
If \code{NULL}, all events are returned.}

\item{survey}{Logical. Specifies whether or not to export the survey identifier field
(e.g., "redcap_survey_identifier") or survey timestamp fields
(e.g., form_name+"_timestamp") when surveys are utilized in the project.
If you do not pass in this flag, it will default to "false". If set to
"true", it will return the redcap_survey_identifier field and also the
survey timestamp field for a particular survey when at least
one field from that survey is being exported. NOTE: If the survey
identifier field or survey timestamp fields are imported via API data
import, they will simply be ignored since they are not real fields in
the project but rather are pseudo-fields.}

\item{dag}{Logical. Specifies whether or not to export the "redcap_data_access_group"
field when data access groups are utilized in the project. If you do not
pass in this flag, it will default to "false". NOTE: This flag is only
viable if the user whose token is being used to make the API request is
\emph{not} in a data access group. If the user is in a group, then this
flag will revert to its default value.}

\item{form_complete_auto}{Logical. If \code{fields} are passed,
REDCap does not return form complete fields unless specifically requested.
However, if \code{TRUE}, the \code{[form]_complete} fields for any form
from which at least one variable is requested will automatically be
retrieved.}

\item{format}{Logical.  Determines whether the data will be formatted with
\code{rc_format} (Default = FALSE)}

\item{...}{Additional arguments to be passed to \code{rc_format}}

\item{strip}{Logical. If \code{TRUE}, empty rows and columns will be removed from
record_data. See \code{rc_strip} for more information or call seperately for more
options.}

\item{colClasses}{A (named) vector of colum classes passed to
\code{\link[utils]{read.csv}} calls.
Useful to force the interpretation of a column in a specific type and
avoid an unexpected recast.}

\item{batch.size}{Integer.  Specifies the number of subjects to be included
in each batch of a batched export.  Non-positive numbers export the
entire project in a single batch. Batching the export may be beneficial
to prevent tying up smaller servers.  See details for more explanation.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcap_error}}}
}
\description{
Exports records from a REDCap Database, allowing for
subsets of subjects, fields, records, and events. By default, all
records will be exported. If a report ID is supplied, only data
from that report will be exported.
}
\details{
A record of exports through the API is recorded in the Logging section
of the project.

It is unnecessary to include "redcap_event_name" or the "redcap_repeat" variables
in the fields argument. These fields are automatically exported for any
longitudinal database. If the user does include them in the fields argument,
they are removed quietly in the parameter checks.

A 'batched' export is one where the export is performed over a series of
API calls rather than one large call.  For large projects on small servers,
this may prevent a single user from tying up the server and forcing others
to wait on a larger job.  The batched export is performed by first
calling the API to export the subject identifier field (the first field
in the meta data).  The unique ID's are then assigned a batch number with
no more than \code{batch.size} ID's in any single batch.  The batches are
exported from the API and stacked together.

In longitudinal projects, \code{batch.size} may not necessarily be the
number of records exported in each batch.  If \code{batch.size} is 10 and
there are four records per patient, each batch will consist of 40 records.
Thus, if you are concerned about tying up the server with a large,
longitudinal project, it would be prudent to use a smaller batch size.

Note about export rights (6.0.0+): Please be aware that Data Export user rights will be
applied to this API request. For example, if you have "No Access" data export rights
in the project, then the API data export will fail and return an error. And if you
have "De-Identified" or "Remove all tagged Identifier fields" data export rights,
then some data fields \emph{might} be removed and filtered out of the data set returned
from the API. To make sure that no data is unnecessarily filtered out of your API
request, you should have "Full Data Set" export rights in the project.

REDCap Version:= 6.0.0

Deidentified Batched Calls:
Batched calls to the API are not a feature of the REDCap API, but may be imposed
by making multiple calls to the API.  The process of batching the export requires
that an initial call be made to the API to retrieve only the record IDs.  The
list of IDs is then broken into chunks, each about the size of \code{batch.size}.
The batched calls then force the \code{records} argument in each call.

When a user's permissions require a de-identified data export, a batched call
should be expected to fail.  This is because, upon export, REDCap will hash the
identifiers.  When R attempts to pass the hashed identifiers back to REDCap,
REDCap will try to match the hashed identifiers to the unhashed identifiers in the
database.  No matches will be found, and the export will fail.

Users who are exporting de-identified data will have to settle for using unbatched
calls to the API (ie, \code{batch.size = -1})
}
\references{
Please refer to your institution's API documentation.

Additional details on API parameters are found on the package wiki at
\url{https://github.com/nutterb/redcapAPI/wiki/REDCap-API-Parameters}

This functionality was originally developed by Jeffrey Horner in the \code{redcap} package.
\url{https://github.com/vubiostat/redcap}

See also \code{read_redcap_oneshot} in the \code{REDCapR} package by Will Beasley.
\url{https://github.com/OuhscBbmc/REDCapR}

Borrowed code from http://stackoverflow.com/a/8099431/1017276 to
create a list of arbitrary length.
}
\author{
Jeffrey Horner

Marcus Lehr
}
