% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_exportReports.R
\name{rc_exportReports}
\alias{rc_exportReports}
\title{Export Reports from a REDCap Database}
\usage{
rc_exportReports(
  report_id,
  url = getOption("redcap_bundle")$redcap_url,
  token = getOption("redcap_token"),
  meta_data = getOption("redcap_bundle")$meta_data,
  format = TRUE,
  ...,
  error_handling = getOption("redcap_error_handling")
)
}
\arguments{
\item{report_id}{Integer.  Gives the report id of the desired report. 
This is located on the Report Builder page of the user interface on REDCap.}

\item{url}{A url address to connect to the REDCap API}

\item{token}{A REDCap API token}

\item{meta_data}{REDCap project metadata (aka data dictionary). By default, 
$meta_data is expected in a REDCap bundle object, as created by \code{rc_setup}.
Otherwise, a data.frame containing the metadata must be supplied.}

\item{format}{Logical.  Determines whether the data will be formatted with
\code{rc_formatRecords} (Default = TRUE)}

\item{...}{Additional arguments to be passed to \code{rc_formatRecords}}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcap_error}}}
}
\description{
Exports reports from a REDCap Database and formats data if requested
}
\details{
A record of exports through the API is recorded in the Logging section of 
the project.

Reports are exported based on their id number, which can be looked up in 
the Reports page of a project
}
\section{REDCap API Documentation (6.5.0)}{

This function allows you to export the data set of a report created on a project's 
"Data Exports, Reports, and Stats" page.

Note about export rights (6.0.0+): Please be aware that Data Export user rights will be 
applied to this API request. For example, if you have "No Access" data export rights 
in the project, then the API report export will fail and return an error. And if you 
have "De-Identified" or "Remove all tagged Identifier fields" data export rights, 
then some data fields *might* be removed and filtered out of the data set returned 
from the API. To make sure that no data is unnecessarily filtered out of your API 
request, you should have "Full Data Set" export rights in the project.
}

\section{REDCap Version}{

6.0.0+
}

\section{Known REDCap Limitations}{

None
}

\author{
Benjamin Nutter

Marcus Lehr
}
