% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_findMissing.R
\name{rc_findMissing}
\alias{rc_findMissing}
\title{Find missing record data}
\usage{
rc_findMissing(
  records,
  completionField,
  meta_data = getOption("redcap_bundle")$meta_data,
  events = getOption("redcap_bundle")$events
)
}
\arguments{
\item{records}{A raw data export from REDCap.}

\item{completionField}{The REDCap variable which indicates whether or not a subject
has completed the study. This should be indicated by a 'Yes' or a '1' (i.e. a Yes/No
field in REDCap).}

\item{meta_data}{REDCap project metadata (aka data dictionary). By default, 
$meta_data is expected in a REDCap bundle object, as created by \code{rc_setup}.
Otherwise, a data.frame containing the metadata must be supplied.}

\item{events}{REDCap project event metadata. By default, $events is expected in
the REDCap bundle, as created by \code{rc_setup}. Otherwise, a data.frame containing
event data must be supplied.}
}
\description{
Finds data points which are ostenibly missing from Redcap records data.
}
\details{
For each event/variable combination, if a value exists for at least one
record then it is expected to exist for all records. If an entire event is empty for
a given record, none of the variables will be flagged unless there is data for that
record in the following event. Additionally, if the subject is marked as having
completed the study (i.e. \code{completionField == 'Yes'|1}), then empty events for
that subject will not be discarded.

The logic of this function does not extend to checkbox variables or repeat instruments.
If present in the record data, they will be dropped.
}
\author{
Marcus Lehr
}
