% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_format.R
\name{rc_format}
\alias{rc_format}
\title{Format records data}
\usage{
rc_format(
  record_data,
  data_dict = getOption("redcap_bundle")$data_dict,
  event_data = getOption("redcap_bundle")$event_data,
  factor_labels = TRUE,
  col_labels = TRUE,
  dates = TRUE,
  event_labels = TRUE,
  strip = FALSE
)
}
\arguments{
\item{record_data}{Dataframe. Record data export from REDCap}

\item{data_dict}{Dataframe. REDCap project data data_dictionary. By default,
this will be fetched from the REDCap bundle option, as created by \code{rc_bundle}.
Otherwise, a data.frame containing the project data dictionary must be supplied.}

\item{event_data}{Dataframe. REDCap event data. By default, this will be fetched
from the REDCap bundle option, as created by \code{rc_bundle}.}

\item{factor_labels}{Logical.  Determines whether categorical fields (eg radio,
dropdown, checkbox) will be formatted as numeric codes or labels. Note that
for checkbox fields, 0 (representing an unchecked box) will be replaced with NA.}

\item{col_labels}{Logical.  Determines if the field labels are applied to the
dataframe columns.}

\item{dates}{Logical. Determines if date variables are converted to POSIXct
format during the download.}

\item{event_labels}{Logical. Determines if event labels are applied to redcap_event_name
column (Default = \code{TRUE}) or left as unique event names.}

\item{strip}{Logical. If \code{FALSE}, empty rows and columns will be removed from
record_data. See \code{rc_strip} for more information or call separately for more
options.}
}
\description{
Uses REDCap project metadata to format records data.
}
\details{
This function takes raw REDCap data and adds column labels, converts
columns to numeric/character/factor as appropriate, and applies factor and
checkbox labels. Formatting details of the returned dataframe can be found
via attributes(record_data)$redcap_formatting.
}
\author{
Marcus Lehr

Benjamin Nutter
}
