% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_import.R
\name{rc_import}
\alias{rc_import}
\title{Import Records to a REDCap Database}
\usage{
rc_import(
  record_data,
  url = getOption("redcap_bundle")$redcap_url,
  token = getOption("redcap_token"),
  data_dict = getOption("redcap_bundle")$data_dict,
  overwriteBehavior = "normal",
  returnContent = "count",
  returnData = FALSE,
  logfile = "",
  batch_size = -1
)
}
\arguments{
\item{record_data}{A \code{data.frame} to be imported to the REDCap project.}

\item{url}{A url address to connect to the REDCap API}

\item{token}{Path to a text file containing your REDCap API token}

\item{data_dict}{Dataframe. A REDCap project data dictionary. By default,
$data_dict is expected in the REDCap bundle option, as created by
\code{rc_bundle}.}

\item{overwriteBehavior}{Character string.  'normal' prevents existing
data in REDCap from being altered. Note that when set to 'overwrite',
blank values (NAs) will replace existing data.}

\item{returnContent}{Character string.  'count' returns the number of
records imported; 'ids' returns a list of record ids imported;
'nothing' returns no message.}

\item{returnData}{Logical.  Prevents the REDCap import and instead
returns the data frame that would have been given
for import.  This is sometimes helpful if the API import fails without
providing an informative message. The data frame can be written to a csv
and uploaded using the interactive tools to troubleshoot the
problem.  Please shoot me an e-mail if you find errors I haven't
accounted for.}

\item{logfile}{An optional filepath (preferably .txt) in which to print the
log of errors and warnings about the data.
If \code{""}, the log is printed to the console.}

\item{batch_size}{Specifies size of batches.  A negative value
indicates no batching.}
}
\description{
Imports records from a \code{data.frame} to a REDCap Database
}
\details{
A record of imports through the API is recorded in the Logging section
of the project.

\code{rc_import} prevents the most common import errors by testing the
data before attempting the import.  Namely
\enumerate{
\item Check that all variables in \code{record_data} exist in the REDCap data data_dictionary.
\item Check that the study id variable exists
\item Force the study id variable to the first position in the data frame (with a warning)
\item Remove calculated fields (with a warning)
\item Verify that REDCap date fields are represented in the data frame as
either character, POSIXct, or Date class objects.
\item Determine if values are within their specified validation limits.
}

See the documentation for \code{\link{validateImport}} for detailed
explanations of the validation.
}
\references{
\url{http://stackoverflow.com/questions/12393004/parsing-back-to-messy-api-strcuture/12435389#12435389}\cr
\url{https://github.com/etb/my-R-code/blob/master/R-pull-and-push-from-and-to-REDCap.R}\cr
See also the REDCap API documentation
Please refer to your institution's API documentation.

Additional details on API parameters are found on the package wiki at
\url{https://github.com/nutterb/redcapAPI/wiki/REDCap-API-Parameters}
}
\seealso{
\code{\link{validateImport}}
}
\author{
Benjamin Nutter\cr
with thanks to Josh O'Brien and etb (see references)

Marcus Lehr
}
