% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_logicCheck.R
\name{rc_logicCheck}
\alias{rc_logicCheck}
\title{Check branching logic for errors}
\usage{
rc_logicCheck(
  meta_data = getOption("redcap_bundle")$meta_data,
  events = getOption("redcap_bundle")$events
)
}
\arguments{
\item{meta_data}{REDCap project metadata (aka data dictionary). By default, 
$meta_data is expected in a REDCap bundle object, as created by \code{rc_setup}.
Otherwise, a data.frame containing the metadata must be supplied.}
}
\description{
Check branching logic for common errors
}
\details{
This function will verify that all event and variable names used in the
branching logic exist in the project metadata. It will also warn of faulty logic
concerning \code{[event-name] = 'event_name'} conditions.
}
\author{
Marcus Lehr
}
