% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_missing.R
\name{rc_missing}
\alias{rc_missing}
\title{Find missing record data}
\usage{
rc_missing(
  record_data,
  completion_field = NULL,
  repeats = NULL,
  table = FALSE,
  bundle = getOption("redcap_bundle"),
  data_dict = getOption("redcap_bundle")$data_dict,
  event_data = getOption("redcap_bundle")$event_data,
  form_perm = getOption("redcap_bundle")$form_perm
)
}
\arguments{
\item{record_data}{Record data export from REDCap}

\item{completion_field}{Character. The REDCap variable which indicates whether or not a subject
has completed the study. This should be indicated by a 'Yes' or a '1' (i.e. a Yes/No
field in REDCap). If not provided, some missing data may be lost.}

\item{repeats}{Character. A vector of repeat instrument names which should be kept. This should
only be used for repeat instruments which are expected to have a consistent number of occurances
per each event/ID combination. All repeat instruments not supplied here will be removed to avoid
excessive false positives.}

\item{table}{Logical. If \code{TRUE} a list will be returned containing the missing dataframe
and two missing tables. The tables contain the percent of missing record_idor variable counts.
To view formatted tables, use square bracket indexing- e.g. data[\link{'IDs_table'}]}

\item{bundle}{List. A project metadata bundle as created by \code{rc_setup}. By default, the
bundle is expected in the option "redcap_bundle".}

\item{data_dict}{Dataframe. REDCap project data data_dictionary. By default,
$data_dict is expected in a REDCap bundle object, as created by \code{rc_setup}. Otherwise, an
equivalent data.frame containing the project data dictionary must be supplied.}

\item{event_data}{Dataframe. REDCap project event metadata. By default, $event_data is expected in
the REDCap bundle, as created by \code{rc_setup}. Otherwise, an equivalent data.frame containing
event data must be supplied.}

\item{form_perm}{Dataframe. User access permissions for REDCap forms. By default, $form_perm
is expected in the REDCap bundle, as created by \code{rc_setup}. Otherwise, an equivalent
data.frame containing form permission data must be supplied.}
}
\description{
Finds data points which are ostenibly missing from Redcap records data.

For each event/variable combination, if a value exists for at least one
record then it is expected to exist for all records. Any event or variable which is empty within
the entire dataframe will inherently be dropped. If an event has no values for a given record,
none of the variables will be flagged unless there is data for that record in the following event.
Additionally, if the subject is marked as having completed the study (i.e. \code{completion_field 
== 'Yes'|1}), then empty events for that subject will not be discarded.

The logic of this function does not extend to repeat instruments or checkbox, calculated,
or hidden fields. If present in the record data, they will be dropped. Additionally, this function
does not yet support branching logic. It will be implemented in a future version.
}
\author{
Marcus Lehr
}
