% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_missing_plot.R
\name{rc_missing_plot}
\alias{rc_missing_plot}
\title{Plot data from rc_missing}
\usage{
rc_missing_plot(missing_data)
}
\arguments{
\item{missing_data}{Dataframe. A frame of long format missing data, as
created by \code{rc_missing}.}
}
\description{
Creates a pie chart for each variable passed from
\code{rc_missing} and returns list containing the plots. \emph{NOTE} The
data passed from \code{rc_missing} must be in long format. These
plots are useful for investigating the structure of missing data.
}
\author{
Marcus Lehr
}
