% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_outliers_plot.R
\name{rc_outliers_plot}
\alias{rc_outliers_plot}
\title{Plot data from rc_outliers}
\usage{
rc_outliers_plot(
  outlier_data,
  variable_var = "variable",
  value_var = "value",
  outlier_var = "outlier",
  sex_var = NA,
  id_field = getOption("redcap_bundle")$id_field
)
}
\arguments{
\item{outlier_data}{Dataframe. A long format data frame, as created by
\code{rc_outliers} using the argument \code{unfiltered = T}.}

\item{variable_var}{Character. Name of the column containing variable names.}

\item{value_var}{Character. Name of the column containing numeric data to
be plotted.}

\item{outlier_var}{Character. Name of the column containing logical (T/F)
data indicating which values are outliers.}

\item{sex_var}{String. Name of variable indicating the sex of subjects. If
included, a list object containing separate plots for each sex will be returned.}

\item{id_field}{Character. Field name corresponding to the 'record_id' field.}
}
\description{
Creates a grid of plots for each variable passed from
\code{rc_outliers}. \emph{NOTE} In order for this function to be useful, the
argument \code{unfiltered} should be set to \code{FALSE} when running
\code{rc_outliers}. Outliers are colored in red and horizontal indicator
lines for +/- 3 standard deviations are included. If a \code{sex_var} is
included, a list object will be returned containing a plot for each sex.
}
\author{
Marcus Lehr
}
