% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_plot.R
\name{rc_plot}
\alias{rc_plot}
\title{Plot long format data}
\usage{
rc_plot(
  molten_data,
  title = NA,
  plot_type = "standard",
  outlier_var = "outlier",
  wrap_var = "variable",
  y = "value",
  x = getOption("redcap_bundle")$id_field
)
}
\arguments{
\item{molten_data}{Dataframe. A long format data frame, as created by
\code{rc_outliers} using the argument \code{unfiltered = T}.}

\item{title}{Character. Title of the resulting plot.}

\item{plot_type}{Character. One of 'standard', 'qq', or 'hist'}

\item{outlier_var}{Character. Name of the column containing logical (T/F)
data indicating which values are outliers. Change to NA when not in use.}

\item{wrap_var}{Character. Name of the column containing variable names.}

\item{y}{Character. Name of the column containing numeric data to
be plotted.}

\item{x}{Character. Field name corresponding to the 'record_id' field.}
}
\description{
Creates a grid of plots for each variable passed from
\code{rc_outliers}. \emph{NOTE} In order for this function to be useful, the
argument \code{unfiltered} should be set to \code{FALSE} when running
\code{rc_outliers}. Outliers are colored in red and horizontal indicator
lines for +/- 3 standard deviations are included. If a \code{sex_var} is
included, a list object will be returned containing a plot for each sex.
}
\author{
Marcus Lehr
}
