% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_plot_outliers.R
\name{rc_plot_outliers}
\alias{rc_plot_outliers}
\title{Plot data from rc_outliers}
\usage{
rc_plot_outliers(
  outlier_data,
  sex_var = NA,
  id_field = getOption("redcap_bundle")$id_field
)
}
\arguments{
\item{outlier_data}{Dataframe. A long format data frame, as created by
\code{rc_outliers} using the argument \code{unfiltered = T}.}

\item{sex_var}{String. Name of variable indicating the sex of subjects. If
included, a list object containing separate plots for each sex will be returned.}

\item{id_field}{Character. Field name corresponding to the 'record_id' field.}
}
\description{
Creates a grid of plots for each variable passed from
\code{rc_outliers}. \emph{NOTE} In order for this function to be useful, the
argument \code{unfiltered} should be set to \code{FALSE} when running
\code{rc_outliers}. Outliers are colored in red and horizontal indicator
lines for +/- 3 standard deviations are included. If a \code{sex_var} is
included, a list object will be returned containing a plot for each sex.
}
\author{
Marcus Lehr
}
