% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_pool.R
\name{rc_pool}
\alias{rc_pool}
\title{Pools columns for aggregated analysis of fields with like data}
\usage{
rc_pool(
  record_data,
  var_roots = NULL,
  fields_list = NULL,
  long_format = FALSE,
  id_field = getOption("redcap_bundle")$id_field
)
}
\arguments{
\item{record_data}{Dataframe. Records data export from REDCap. For the
purposes of this function, only quantitative data will be kept.}

\item{var_roots}{Character. Strings to search for within column names of
record_data. For each variable root provided, all column names containing the
root will be pooled into a single column. In order to prevent inappropriate
matches, the search space will be restricted to columns which appear to contain
numerical data.}

\item{fields_list}{List. A list in the format \code{list(new_column = c("old","column","names"))}.
Unlike \code{var_roots}, the column names provided here will be matched exactly.
In addition, if both \code{var_roots} and \code{fields_list} are provided,
\code{fields_list} will be applied first.}

\item{long_format}{Logical. Determines whether the returned dataframe will
be in long or wide format. Default is \code{FALSE}. This option is useful for more
aggressive pooling- particularly when there are same-row data points that will be
lost in wide format.}

\item{id_field}{Character. Field name corresponding to the 'record_id' field.}
}
\description{
For each variable root provided, all column names in the
record_data containing that root will be pooled into a single column and
appended to the end of the dataframe. Roots provided via \code{var_roots}
will only be matched to columns containing numeric data to reduce false matches.
An attribute describing the columns pooled will be appended to the dataframe,
which can be accessed via the command

Additionally, exact (i.e. full name) matching can be performed with the
\code{fields_list} argument. Fields provided in this argument will be searched
for in all columns. If both arguments are provided, \code{fields_list} will be
applied first.

Furthermore, if the columns selected to be pooled contain more than one
data point per row, the first data point will be used. In this case, pooling is
likely inappropriate and the pooled columns should be reviewed. To see which columns
have been pooled, run the command \code{attributes([YOUR_DATA_FRAME])$pooled_vars}
on the returned dataframe. If for some reason pooling is still desirable and all data
points should be kept, use \code{long_format = TRUE} for more aggressive pooling.
}
\details{
The intention of this function is to correct for inefficient
REDCap project design where the same data measurement has been assigned to
multiple variables. For example, if the variables "visit_1_weight" and
"visit_2_weight" have been used to collect weight at different visits rather
than re-using the same variable, they can be pooled into a single column using
the \code{var_root} "weight". This is often desirable for analysis.
}
\author{
Marcus Lehr
}
