% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_repDiff.R
\name{rc_repDiff}
\alias{rc_repDiff}
\title{Report differences between technical replicates}
\usage{
rc_repDiff(records, varRoots, meta_data = getOption("redcap_bundle")$meta_data)
}
\arguments{
\item{records}{A raw data export from REDCap.}

\item{varRoots}{A vector of root names which are common and unique to all of the associated
technical replicates.}

\item{meta_data}{REDCap project metadata (aka data dictionary). By default, 
$meta_data is expected in a REDCap bundle object, as created by \code{rc_setup}.
Otherwise, a data.frame containing the metadata must be supplied.}
}
\description{
Take differences for each sequential pair of replicate variables
}
\details{
For each variable root supplied, all variables containing the root in the
record data will be collected and a new column containing difference values for
each sequential pair will be created.
}
\author{
Marcus Lehr
}
