% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_strip.R
\name{rc_strip}
\alias{rc_strip}
\title{Remove empty rows and columns from a REDCap dataframe}
\usage{
rc_strip(
  record_data,
  columns = TRUE,
  rows = TRUE,
  id_field = getOption("redcap_bundle")$id_field
)
}
\arguments{
\item{record_data}{Dataframe. Record data exported from REDCap}

\item{columns}{Logical. When \code{TRUE}, empty columns will
be removed.}

\item{rows}{Logical. When \code{TRUE}, empty rows will be removed.}

\item{id_field}{Character. The name of the record_id field for your REDCap project.}
}
\description{
Removes any columns which contain only \code{NA}. Likewise,
rows are removed in a similar fashion but any REDCap factor fields
(record_id, redcap event/repeat) are ignored.
}
\author{
Marcus Lehr
}
