% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_to_options.R
\name{rc_to_options}
\alias{rc_to_options}
\title{Upload local bundle or token to options}
\usage{
rc_to_options(bundle = NULL, token = NULL)
}
\arguments{
\item{bundle}{A REDCap bundle object}

\item{token}{Path to a text file containing your REDCap API token}
}
\description{
Uploads a local REDCap bundle object to the R session options.
}
\author{
Marcus Lehr
}
