% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcTools-package.R
\docType{package}
\name{RcTools-package}
\alias{RcTools}
\alias{RcTools-package}
\title{RcTools: A Collection of Report Card Tools}
\description{
This package makes a collection of Report Card functions available to the technical team of the Northern Three Partnerships. Tools are mostly focused on styling tabular outputs for inclusion in the Technical Report, this includes conditional formatting, assigning grades, and cell colouring.
}
\author{
\strong{Maintainer}: Adam Shand \email{adamshand22@gmail.com}

}
\keyword{internal}
