% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_splitter.R
\name{row_splitter}
\alias{row_splitter}
\title{Split a Table into N rows}
\usage{
row_splitter(df, RowCount = 1500, NameVec = NULL)
}
\arguments{
\item{df}{A single dataframe or a list of dataframes}

\item{RowCount}{The number of rows per split dataframe. Defaults to 1500}

\item{NameVec}{A single dataframe name or list of dataframe names, of equal length to the number of dataframes provided under df.
Defaults to NULL and will fill dataframe names with the generic "table_n"}
}
\value{
A list of named dataframes
}
\description{
Split a Table into N rows
}
\examples{

#a simple example with a single dataframe
df <- data.frame(Name = rep("example_name", 10000), Value = rep(1, 10000))
split_dfs <- row_splitter(df, 1000, "my_df")

#an example with several dataframes
a_2025 <- data.frame(Name = rep("A", 10000), Year = rep(2025, 10000), Value = rep(1, 10000))
a_2024 <- data.frame(Name = rep("A", 10000), Year = rep(2024, 10000), Value = rep(1, 10000))
b_2025 <- data.frame(Name = rep("B", 10000), Year = rep(2025, 10000), Value = rep(1, 10000))
b_2024 <- data.frame(Name = rep("B", 10000), Year = rep(2024, 10000), Value = rep(1, 10000))

df_list <- list(a_2025, a_2024, b_2025, b_2024)
df_names <- list("a_2025"," a_2024", "b_2025", "b_2024")

split_dfs <- row_splitter(df_list, 1000, df_names)

}
