% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_to_grade.R
\name{score_to_grade}
\alias{score_to_grade}
\title{Function that converts scores to grades}
\usage{
score_to_grade(df, score_cols)
}
\arguments{
\item{df}{Dataframe. The table that contains the score(s) you are interested in.}

\item{score_cols}{Numeric column(s). Any number of numeric columns that contain scores. Values are expected to
range from 0 to 100}
}
\value{
The original dataframe with new grade columns added. Names are inherited from the targeted columns.
}
\description{
Function that converts scores to grades
}
\examples{
\dontrun{
x <- data.frame(
 Basin = c("Black", "Ross", "Haughton", "Suttor"),
 ScoreA = c(51, 76, 27, 98),
 ScoreB = c(7, 42, 89, 63),
 ScoreC = c(15, 34, 56, 21)
)
x <- score_to_grade(x, c(ScoreA, ScoreB, ScoreC))
  
}
}
