% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_change_over_time.R
\name{sf_change_over_time}
\alias{sf_change_over_time}
\title{Compare how a dataset changes over time}
\usage{
sf_change_over_time(sf, Col, StartYear, EndYear, Resolution = 1000)
}
\arguments{
\item{sf}{An sf object. Generally the object should contain polygons geometries, and must contain at least 2 years of data.}

\item{Col}{Character string. The name of the column from which to extract information about the geometry}

\item{StartYear}{Numeric string. The first year of data to use. Must exist within the sf object.}

\item{EndYear}{Numeric string. The second year of data to use. Must exist within the sf object.}

\item{Resolution}{Numeric string. The number of grid cells to create along the x and y planes - will for a square grid. Defaults to 1000}
}
\value{
An sf object
}
\description{
Compare how a dataset changes over time
}
\examples{

polygons <- list(
  rbind(c(0,0), c(1,0), c(1,1), c(0,1), c(0,0)), 
  rbind(c(1,0), c(2,0), c(2,1), c(1,1), c(1,0)),
  rbind(c(0,1), c(1,1), c(1,2), c(0,2), c(0,1)),
  rbind(c(1,1), c(2,1), c(2,2), c(1,2), c(1,1)))

polygons <- sf::st_sfc(lapply(polygons, function(x) sf::st_polygon(list(x))))

poly_sf <- dplyr::bind_rows(
  sf::st_sf(Year = 2000, Id = c("a", "b", "c", "d"), geometry = polygons, crs = 4326), 
  sf::st_sf(Year = 2020, Id = rep("a",4), geometry = polygons, crs = 4326))

change <- sf_change_over_time(poly_sf, "Id", 2000, 2020)

tmap::tm_shape(change) + tmap::tm_polygons(fill = "ValueChange")

}
