% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl4.R
\docType{data}
\name{tbl4}
\alias{tbl4}
\title{Hypothetical Technical Report Data 4}
\format{
\subsection{\code{tbl4}}{

A dataframe with 13 rows and 38 columns:
\describe{
\item{Basin}{A string column containing the Basin in which the site is found}
\item{SiteNumber}{A numeric column containing the site number}
\item{X1}{A numeric column containing either 1 or 0, indicating whether the species was observed}
\item{X2}{A numeric column containing either 1 or 0, indicating whether the species was observed}
\item{...}{etc.}
...
}
}
}
\source{
NA - this is a completely made up dataset.
}
\usage{
tbl4
}
\description{
An example of presence/absence data for a hypothetical technical report,
specifically focusing on fish observations
}
\keyword{datasets}
