% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl5.R
\docType{data}
\name{tbl5}
\alias{tbl5}
\title{Table Containing Testing Values}
\format{
\subsection{\code{tbl5}}{

A dataframe with 14 rows and 10 columns:
\describe{
\item{wqo}{A numeric column containing two different water quality objective value options}
\item{sf}{A numeric column containing two different scaling factor options}
\item{eightieth}{A numeric column containing two different eightieth value options}
\item{twentieth}{A numeric column containing two different twentieth value options}
\item{water_quality_indicator}{A string column defining the type of water quality indicator scored for the row}
\item{fish_indicator}{A string column defining the type of fish indicator scored for the row}
\item{water_quality_test}{A numeric column containing a range of possible water quality values}
\item{wetland_test}{A numeric column containing a range of possible wetland values}
\item{riparian_test}{A numeric column containing a range of possible riparian values}
\item{fish_test}{A numeric column containing a range of possible fish values}
...
}
}
}
\source{
NA - this is a completely made up dataset.
}
\usage{
tbl5
}
\description{
A table containing a range of values to perform tests on the scoring and grading functions
}
\keyword{datasets}
