% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_to_score.R
\name{value_to_score}
\alias{value_to_score}
\title{Function that converts values to score}
\usage{
value_to_score(
  value,
  value_type,
  indicator = NULL,
  water_type = NULL,
  wqo = NULL,
  sf = NULL,
  eightieth = NULL,
  twentieth = NULL
)
}
\arguments{
\item{value}{Numeric Vector. Any vector of numbers that are values.}

\item{value_type}{String. Defines the type of values that will be scored. One of: "water quality", "wetlands", "mangroves and saltmarsh"}

\item{indicator}{String Vector. Defines the indicator that will be scored. Only relevant for water quality and fish. Provide as a vector of equal length to values.}

\item{water_type}{String. Defines the type of water quality scoring to use. One of "Freshwater", "Estuarine", or "Marine"}

\item{wqo}{Integer. The water quality objective. Only relevant for water quality}

\item{sf}{Integer.  The scaling factor. Only relevant for freshwater and estuarine water quality}

\item{eightieth}{Integer. The eightieth percentile. Only relevant for freshwater and estuarine water quality}

\item{twentieth}{Integer. The twentieth percentil. Only relevant for freshwater and estuarine water quality}
}
\value{
A vector. Code suggesting provided for imbedding the vector back into a df
}
\description{
Function that converts values to score
}
\examples{
\dontrun{
x <- data.frame(
 Basin = c("Black", "Ross", "Haughton", "Suttor"),
 Value = c(, 76, 27, 98)
)
x <- x |> 
  mutate(across(
    Score,
    .names = "{.col}_Score",
    ~ value_to_score(.)))
 
}
}
