% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_scores.R
\name{weight_score}
\alias{weight_score}
\title{Function that weights scores based on predetermined weighting values}
\usage{
weight_score(df, score, weighting)
}
\arguments{
\item{df}{Dataframe. The table that contains the value(s) you are interested in.}

\item{score}{Numeric column(s). Any number of numeric columns that contain scores to be weighted.}

\item{weighting}{Numeric column. A singular numeric column that contains weighting values, must be the same length
as the score column(s). Weights should be decimal (i.e. a 25\% weighting should be written as 0.25).}
}
\value{
The original dataframe with new weighted scores added. Names are inherited from the targeted columns.
}
\description{
Function that weights scores based on predetermined weighting values
}
\examples{
\dontrun{
x <- data.frame(
 Basin = c("Black", "Ross", "Haughton", "Suttor"),
 Indicator = c("DIN", "DIN", "Low_DO", "Low_DO"),
 Score = c(100, 55, 67, 18),
 Weight = c(0.25, 0.5, 0.3, 0.2)
)

x <- x |> 
  weight_score(
    score = Score,
    weighting = Weight
  )
 
}
}
