% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_letter_to_score.R
\name{bind_letter_to_score}
\alias{bind_letter_to_score}
\title{Helper to Extract A Letter Grade From Scores}
\usage{
bind_letter_to_score(df, columns, rows)
}
\arguments{
\item{df}{A dataframe/table/spreadsheet}

\item{columns}{Range. The column indices to be targeted.}

\item{rows}{Range. The row indicies to be targeted.}
}
\value{
A dataframe with letters attached to scores
}
\description{
Helper to Extract A Letter Grade From Scores
}
\examples{
\dontrun{
df <- data.frame(
 "Areas" = c("A", "B", "C", "D", "E", "F"),
 "Score1" = c(0, 20, 40, 60, 80, 100),
 "Score1" = c(0, 20, 40, 60, 80, 100),
 "Other" = c("Z", "Y", "X", "W", "V", "U")
)
column_indicies <- 2:3
row_indicies <- 1:4
bind_letter_to_score(df, column_indicies, row_indicies)
}
}
