% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_base_map.R, R/dem_pre_processing.R
\name{dem_pre_processing}
\alias{dem_pre_processing}
\title{Create the base map elements for a DEM map}
\usage{
dem_pre_processing(RawPath, CropPath, OutputName, CropObj, Reload, Overwrite)

dem_pre_processing(RawPath, CropPath, OutputName, CropObj, Reload, Overwrite)
}
\arguments{
\item{RawPath}{The path to where to find the raw, unprocessed data}

\item{CropPath}{The path where the function should save the processed data}

\item{OutputName}{The name of the file when it is saved}

\item{CropObj}{The object to use for cropping the dataset}

\item{Reload}{Boolean. Do you want to try and reload a prior saved dataset with the same name? Defaults to TRUE.}

\item{Overwrite}{Boolean. Do you want to overwrite the previously saved dataset with the same name? Defaults to FALSE}

\item{sr}{A SpatRaster object, generally produced by the dem_pre_processing() function.}

\item{Resolution}{A numeric vector, either 30 or 100. This should match the resolution of the sr object provided.}

\item{OutputPath}{A character vector that defines the path to the folder where the outputs should be saved}

\item{FileName}{A character vector that uniquely identifies the outputs. Do not include a file type}

\item{Texture}{A character vector. Describes the colour palette to use. One of: “imhof1”, “imhof2”, “imhof3”, “imhof4”, “desert”, “bw”, “unicorn”.}

\item{Zscale}{A numeric vector. Defines ratio between x and y spacing. Defaults to Resolution input (realistic). Decrease
Zscale to exagerate heights}

\item{SeaLevel}{A numeric vector. Defines the elevation of sea level. Defaults to 0m. Decrease or increse to simulate
various sea levels.}
}
\value{
A matrix object and an array object. Both are saved to the path specified and returned to the active environment

A saved .nc file, plus a netcdf object in the active session
}
\description{
Create the base map elements for a DEM map

Preprocess Digital Elevation Model data from AUSSEABED for the GBR
}
\examples{
\dontrun{ #dont run because function takes a long time

dem_cropped <- dem_base_map(
  sr = my_spat_raster,
  Resolution = 30,
  OutputPath = "path/to/output folder/",
  FileNmae = "my_file",
  Reload = FALSE,
  Overwrite = FALSE,
  Texture = "Desert",
  Zscale = 30,
  SeaLevel = 0
)
}

\dontrun{ #dont run because function takes a long time

raw_path <- "path/to/raw folder/"
crop_path <- "path/to/crop folder/"
n3_region <- build_n3_region()

dem_cropped <- dem_pre_processing(raw_path, crop_path, "n3_dem_100m", n3_region, Reload = TRUE, Overwrite = FALSE)
}

}
