% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_pre_processing.R
\name{dem_pre_processing}
\alias{dem_pre_processing}
\title{Preprocess Digital Elevation Model data from AUSSEABED for the GBR}
\usage{
dem_pre_processing(RawPath, CropPath, OutputName, CropObj)
}
\arguments{
\item{RawPath}{The path to where to find the raw, unprocessed data}

\item{CropPath}{The path where the function should save the processed data}

\item{OutputName}{The name of the file when it is saved}

\item{CropObj}{The object to use for cropping the dataset}
}
\value{
A saved .nc file, plus a netcdf object in the active session
}
\description{
Preprocess Digital Elevation Model data from AUSSEABED for the GBR
}
\examples{
\dontrun{ #dont run because function takes a long time

raw_path <- "path/to/raw folder/"
crop_path <- "path/to/crop folder/"
n3_region <- build_n3_region()

dme_cropped <- dem_pre_processing(raw_path, crop_path, "n3_dem_100m", n3_region)
}

}
