% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ereefs_extract.R
\name{ereefs_extract}
\alias{ereefs_extract}
\title{Conditional Formatting with .xlsx Output}
\usage{
ereefs_extract(Region, StartDate, EndDate, Variable, Downsample = 0)
}
\arguments{
\item{Region}{SF Object. An sf object defining the area to extract data from}

\item{StartDate}{POSIX time. Supplied as a string in the format YYYY-MM-DD that defines the first day to extract data from}

\item{EndDate}{POSIX time. Supplied as a string in the format YYYY-MM-DD that defines the last day to extract data from}

\item{Variable}{Character Vector. One or several variables to extract from eReefs, currently accessible: "Turbidity", "Chlorophyll a", "DIN", "NH4", "NO3", "Secchi", "pH", "Wind"}

\item{Downsample}{Integer. A singular value that defines the level of downsampling to apply to the data (reduces data size and processing time, also reduces resolution)}
}
\value{
A NetCDF (stars) object or list of NetCDF (stars) objects
}
\description{
Conditional Formatting with .xlsx Output
}
\examples{
\dontrun{
nc_output <- ereefs_extract(
  Region = an_sf_object,
  StartDate = "2022-03-01",
  EndDate = "2022-03-03",
  Variable = "Turbidity",
  Downsample = 0)
}
}
