% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ereefs_map.R
\name{ereefs_map}
\alias{ereefs_map}
\title{Mapping eReefs Data}
\usage{
ereefs_map(nc)
}
\arguments{
\item{nc}{A netCDF (stars) object, usually produced by the \link{ereefs_extract} and \link{ereefs_reproject} functions}
}
\value{
A tmap object
}
\description{
Mapping eReefs Data
}
\examples{
\dontrun{
m <- ereefs_map(nc)
}
}
