% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_habitat.R
\name{extract_habitat}
\alias{extract_habitat}
\title{Extract Habitat Data from Datasets on File}
\usage{
extract_habitat(RawPath, CropObj, Habitat)
}
\arguments{
\item{RawPath}{Character String. The path to where the full habitat files are found}

\item{CropObj}{Sf Object. The object to be used to crop the files.}

\item{Habitat}{Character String. Either 'RV' or 'MS' for Riparian Vegetation and Mangrove and Saltmarsh respectively. This
tells the function where to save files.}
}
\value{
A table object summarising areas and years. Also crops and saves spatial data, and csv versions.
}
\description{
Extract Habitat Data from Datasets on File
}
\examples{
\dontrun{ #dont run because function takes a long time to process

p1 <- "path/to/folder"

n3_region <- build_n3_region()

all_data <- extract_habitat(p1, n3_region, "RV")
}

}
