% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldToVar.R
\name{fieldToVar}
\alias{fieldToVar}
\title{Convert a REDCap Data Field to an R Vector}
\usage{
fieldToVar(records, data_dict, factor_labels = TRUE, dates = TRUE)
}
\arguments{
\item{records}{A data frame of records returned by \code{rc_export}}

\item{data_dict}{A data frame giving the data data_dictionary, as returned
by \code{exportMetaData}}

\item{factor_labels}{Logical, determines if checkbox, radio button, dropdown and yesno
variables are labelled.}

\item{dates}{Logical, determines if date variables are converted to POSIXct format}
}
\description{
Converts a field exported from REDCap into a valid R vector
}
\details{
This function is called internally by \code{rc_export}. It is not
available to the user.
}
\author{
Jeffrey Horner
}
