% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_cleaning.R
\name{name_cleaning}
\alias{name_cleaning}
\title{A Name Cleaning Function for N3 Tables}
\usage{
name_cleaning(df)
}
\arguments{
\item{df}{Adf object of any dimensions. With or without a geometry column}
}
\value{
A df object of the same dimensions as input
}
\description{
A Name Cleaning Function for N3 Tables
}
\examples{

df <- data.frame(
  "Column 1" = c(1,2,3),
  "column-2" = c(1,2,3),
  "COLUMN    3" = c(1,2,3)
)
df <- name_cleaning(df)

}
