% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qld.R
\docType{data}
\name{qld}
\alias{qld}
\title{Queensland State Border}
\format{
\subsection{\code{qld}}{

An sf object with 883 rows and 21 columns:
\describe{
...
}
}
}
\source{
extracted out of the GISAIMSR Package
}
\usage{
qld
}
\description{
An outline of the Queensland State
}
\keyword{datasets}
