% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_fill.R
\name{rc_fill}
\alias{rc_fill}
\title{Fills in empty rows for categorical data}
\usage{
rc_fill(record_data, ..., group_by = getOption("redcap_bundle")$id_field)
}
\arguments{
\item{record_data}{Dataframe. Record data exported from REDCap}

\item{...}{Variables to be filled. Variable names must be quoted.}

\item{group_by}{Variable(s) to group data by. The project's record_id field
is used by default.}
}
\description{
Takes a data point which only occurs in a single event (row)
and copies that data to all rows for each participant (or other grouping
variable assigned by \code{group_by}). This function is particularly useful
for making filtering operations more straight forward.
}
\author{
Marcus Lehr
}
