% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_logicCheck.R
\name{rc_logicCheck}
\alias{rc_logicCheck}
\title{Check branching logic for errors}
\usage{
rc_logicCheck(bundle)
}
\arguments{
\item{bundle}{A bundle object created by \code{exportBundle} containing project metadata.}
}
\description{
Check branching logic for common errors
}
\details{
This function will verify that all event and variable names used in the
branching logic exist in the project metadata. It will also warn of faulty logic
concerning \code{[event-name] = 'event_name'} conditions.
}
\author{
Marcus Lehr
}
