% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_outliers.R
\name{rc_outliers}
\alias{rc_outliers}
\title{Identify outliers in REDCap records data}
\usage{
rc_outliers(
  record_data,
  sex_var = NA,
  sd_threshold = 3,
  data_dict = getOption("redcap_bundle")$data_dict,
  fields = NULL,
  filtered = FALSE
)
}
\arguments{
\item{record_data}{Dataframe. Records data export from REDCap. For the
purposes of this function, only quantitative data will be kept.}

\item{sex_var}{String. Name of variable indicating the sex of subjects. If
included, variables will be grouped by sex when determining outliers.}

\item{sd_threshold}{Integer. Threshold value for the number of standard
deviations from the mean a value can be before being flagged as an outlier.}

\item{data_dict}{Dataframe. REDCap project data data_dictionary. By default,
$data_dict is expected in the REDCap bundle option, as created by
\code{rc_setup}.}

\item{fields}{Character. A vector of field/variable names to be analyzed
may be passed manually.}

\item{filtered}{Logical. Select whether non-outlier values should be returned.
Default is \code{FLASE}.}
}
\description{
Identifies outliers for each variable, defined as being further
from the mean than the threshold number of standard deviations. Outliers are
then returned as a data.frame.
}
\details{
Unless a vector of variables/field names is passed to the
\code{fields} argument, the fields to be analyzed will be guessed based on
column type. Furthermore, all non-numeric data will be removed before analysis.
If mixed numeric/non-numeric data (e.g. "160 cm") are passed, the first numerical
instance will be extracted from the data. If a sex variable is provided, then
variables will be grouped by sex for outlier analysis.
}
\author{
Marcus Lehr
}
