% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_plot.R
\name{rc_plot}
\alias{rc_plot}
\title{Plot long format data}
\usage{
rc_plot(
  long_data,
  title = NULL,
  plot_type = c("standard", "qq", "hist"),
  outlier_var = NA,
  wrap_var = "variable",
  y = "value",
  x = getOption("redcap_bundle")$id_field,
  legend_position = "none",
  sd_guides = TRUE
)
}
\arguments{
\item{long_data}{Dataframe. A long format data frame, as created by
\code{rc_outliers}. At minimum there should be a record_id column, a
variable name column, and a value column. Wide-to-long conversion functions
include reshape2::melt(), tidyr::gather(), and tidyr::pivot_longer().}

\item{title}{Character. Title of the resulting plot.}

\item{plot_type}{Character. One of 'standard', 'qq', or 'hist'}

\item{outlier_var}{Character. Name of the column containing logical (T/F)
data indicating which values are outliers.}

\item{wrap_var}{Character. Name of the column containing variable names.}

\item{y}{Character. Name of the column containing numeric data to
be plotted.}

\item{x}{Character. Field name corresponding to the 'record_id' field.}

\item{legend_position}{Character. Determines the position of the plot
legend. One of 'none', 'top', 'bottom', 'left' or 'right'.}

\item{sd_guides}{Logical. Determines if standard deviation lines are
applied to plots. Only applies to \code{plot_type = 'standard'}.}
}
\description{
Creates a grid of plots for each variable in long format
data. This function is intended to have data passed from \code{rc_outliers},
plotting the record_id field on the x-axis and values on the y-axis.
Outliers are colored in red and horizontal indicator lines for +/- 3
standard deviations are included by default. Aside from the standard
plots, normal QQ and histograms can also be created using \code{plot_type}.
}
\author{
Marcus Lehr
}
