% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_rep_diff.R
\name{rc_rep_diff}
\alias{rc_rep_diff}
\title{Take differences for each sequential pair of replicate variables}
\usage{
rc_rep_diff(record_data, var_roots)
}
\arguments{
\item{record_data}{A raw data export from REDCap.}

\item{var_roots}{A vector of root names which are common and unique to all of the associated
technical replicates.}
}
\description{
For each variable root supplied, all variables containing the root in the
record data will be collected and a new column containing difference values for
each sequential pair will be created.
}
\author{
Marcus Lehr
}
