% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_repDiff.R
\name{rc_repDiff}
\alias{rc_repDiff}
\title{Report differences between technical replicates}
\usage{
rc_repDiff(records, varRoots, bundle = NULL)
}
\arguments{
\item{records}{A raw data export from REDCap.}

\item{varRoots}{A vector of root names which are common and unique to all of the associated
technical replicates.}

\item{bundle}{A bundle object created by \code{exportBundle} containing project metadata.}
}
\description{
Take differences for each sequential pair of replicate variables
}
\details{
For each variable root supplied, all variables containing the root in the
record data will be collected and a new column containing difference values for
each sequential pair will be created.
}
\author{
Marcus Lehr
}
