% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_strip.R
\name{rc_strip}
\alias{rc_strip}
\title{Remove empty rows and columns from a REDCap dataframe}
\usage{
rc_strip(
  record_data,
  only_rows = FALSE,
  only_columns = FALSE,
  id_field = getOption("redcap_bundle")$id_field
)
}
\arguments{
\item{record_data}{Dataframe. Record data exported from REDCap}

\item{only_rows}{Logical. When \code{TRUE}, only empty rows will
be removed.}

\item{only_columns}{Logical. When \code{TRUE}, only empty columns
will be removed.}
}
\description{
Removes any columns which contain only \code{NA}. Likewise,
rows are removed in a similar fashion but any REDCap factor fields
(record_id, redcap event/repeat) are ignored.
}
\author{
Marcus Lehr
}
