% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcTools-package.R
\docType{package}
\name{RcTools-package}
\alias{RcTools}
\alias{RcTools-package}
\title{RcTools: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
