% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcTools-package.R
\docType{package}
\name{RcTools-package}
\alias{RcTools}
\alias{RcTools-package}
\title{RcTools: A Collection of Report Card Tools}
\description{
"This package makes a collection of Report Card functions available to the technical team of the Northern Three Partnerships. Tools are focused on the final steps of anaylsis and report writing, such as styling tabular outputs for inclusion in the Technical Report, and providing scoring functions and grading functions."
}
\author{
\strong{Maintainer}: Adam Shand \email{adamshand22@gmail.com}

}
\keyword{internal}
