% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_to_grade.R
\name{score_to_grade}
\alias{score_to_grade}
\title{Function that converts scores to grades}
\usage{
score_to_grade(score, grading_type)
}
\arguments{
\item{score}{Numeric Vector. Any vector of numbers that are scores, usually from 0 to 100}
}
\value{
A vector. Code suggesting provided for imbedding the vector back into a df
}
\description{
Function that converts scores to grades
}
\examples{
\dontrun{
x <- data.frame(
 Basin = c("Black", "Ross", "Haughton", "Suttor"),
 Score = c(51, 76, 27, 98)
)
x <- x |> 
  mutate(across(
    Score,
    .names = "{.col}_Grade",
    ~ score_to_grade(.)))
 
}
}
