% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_args.R
\name{validate_args}
\alias{validate_args}
\title{Perform error checking on function arguments}
\usage{
validate_args(
  required = NULL,
  report_id = NULL,
  batch.size = NULL,
  sd_threshold = NULL,
  url = NULL,
  token = NULL,
  id_field = NULL,
  logfile = NULL,
  completion_field = NULL,
  title = NULL,
  outlier_var = NULL,
  wrap_var = NULL,
  y = NULL,
  x = NULL,
  records = NULL,
  fields = NULL,
  forms = NULL,
  events = NULL,
  colClasses = NULL,
  group_by = NULL,
  var_roots = NULL,
  repeats = NULL,
  overwriteBehavior = NULL,
  returnContent = NULL,
  error_handling = NULL,
  plot_type = NULL,
  legend_position = NULL,
  survey = NULL,
  dag = NULL,
  form_complete_auto = NULL,
  format = NULL,
  factors = NULL,
  labels = NULL,
  dates = NULL,
  checkbox_labels = NULL,
  returnData = NULL,
  plot = NULL,
  filtered = NULL,
  long_format = NULL,
  make_repeat = NULL,
  columns = NULL,
  rows = NULL,
  sd_guides = NULL,
  record_data = NULL,
  data_dict = NULL,
  users = NULL,
  form_perm = NULL,
  instruments = NULL,
  arms = NULL,
  mappings = NULL,
  proj_info = NULL,
  long_data = NULL,
  bundle = NULL,
  fields_list = NULL,
  sex_var = NULL
)
}
\arguments{
\item{required}{Character. Vector of argument names which are required by
the calling function.}

\item{report_id}{Numeric, length == 1}

\item{batch.size}{Numeric, length == 1}

\item{sd_threshold}{Numeric, length == 1}

\item{url}{Character, length == 1}

\item{token}{Character, length == 1; 32 alpha-numeric characters}

\item{id_field}{Character, length == 1}

\item{logfile}{Character, length == 1}

\item{completion_field}{Character, length == 1}

\item{title}{Character, length == 1}

\item{outlier_var}{Character, length == 1}

\item{wrap_var}{Character, length == 1}

\item{y}{Character, length == 1}

\item{x}{Character, length == 1}

\item{records}{Character vector}

\item{fields}{Character vector}

\item{forms}{Character vector}

\item{events}{Character vector}

\item{colClasses}{Character vector}

\item{group_by}{Character vector}

\item{var_roots}{Character vector}

\item{repeats}{Character vector}

\item{overwriteBehavior}{Character, defined inputs, length == 1}

\item{returnContent}{Character, defined inputs, length == 1}

\item{error_handling}{Character, defined inputs, length == 1}

\item{plot_type}{Character, defined inputs, length == 1}

\item{legend_position}{Character, defined inputs, length == 1}

\item{survey}{Logical, length == 1}

\item{dag}{Logical, length == 1}

\item{form_complete_auto}{Logical, length == 1}

\item{format}{Logical, length == 1}

\item{factors}{Logical, length == 1}

\item{labels}{Logical, length == 1}

\item{dates}{Logical, length == 1}

\item{checkbox_labels}{Logical, length == 1}

\item{returnData}{Logical, length == 1}

\item{plot}{Logical, length == 1}

\item{filtered}{Logical, length == 1}

\item{long_format}{Logical, length == 1}

\item{make_repeat}{Logical, length == 1}

\item{columns}{Logical, length == 1}

\item{rows}{Logical, length == 1}

\item{sd_guides}{Logical, length == 1}

\item{record_data}{Data.frame; contains record_id and redcap_event_name columns}

\item{data_dict}{Data.frame, ncol == 18}

\item{users}{Data.frame}

\item{form_perm}{Data.frame}

\item{instruments}{Data.frame}

\item{arms}{Data.frame}

\item{mappings}{Data.frame}

\item{proj_info}{Data.frame}

\item{long_data}{Data.frame}

\item{bundle}{List; redcapBundle}

\item{fields_list}{List}

\item{sex_var}{Character (length == 1) or NA}
}
\description{
This will validate all supplied function arguments using the
\code{checkmate} package. All function arguments within the \code{rctools}
package should be contained in this function. To add new arguments, add the
argument to this function, then append the appropriate \code{vars} list in
this function's body.

This is an internal function only.
}
\author{
Marcus Lehr
}
