% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_to_score.R
\name{value_to_score}
\alias{value_to_score}
\title{Function that converts values to score}
\usage{
value_to_score(
  df,
  value,
  value_type,
  water_type = NULL,
  indicator = NULL,
  wqo = NULL,
  sf = NULL,
  eightieth = NULL,
  twentieth = NULL
)
}
\arguments{
\item{df}{Dataframe. The table that contains the value(s) you are interested in.}

\item{value}{Numeric column. One numeric column that contain values to be scored. Values ranges can vary
significantly depending on the indicator.}

\item{value_type}{String. Defines the type of values that will be scored. One of: "Water Quality", "Wetlands",
"Mangroves and Saltmarsh", "Riparian", or "Fish".}

\item{water_type}{String. Defines the type of water quality scoring to use. Only relevant for water quality. One of
"Freshwater", "Estuarine", or "Marine"}

\item{indicator}{String column. Defines the indicator that will be scored. Only relevant for water quality and fish.
Should be sourced from the same df as the value column(s).}

\item{wqo}{Numeric column. The water quality objective associated with each value. Only relevant for water quality.
Should be sourced from the same df as the value column(s).}

\item{sf}{Numeric column. The scaling factor associated with each value. Only relevant for freshwater and estuarine
water quality. Should be sourced from the same df as the value column(s).}

\item{eightieth}{Numeric column. The eightieth percentile associated with each value. Only relevant for freshwater
and estuarine water quality. Should be sourced from the same df as the value column(s).}

\item{twentieth}{Numeric column. The twentieth percentile associated with each value. Only relevant for freshwater
and estuarine water quality. Should be sourced from the same df as the value column(s).}
}
\value{
The original dataframe with new score columns added. Names are inherited from the targeted columns.
}
\description{
Function that converts values to score
}
\examples{

x <- data.frame(
 Basin = c("Black", "Ross", "Haughton", "Suttor"),
 Indicator = c("DIN", "DIN", "Low_DO", "Low_DO"),
 Value = c(0.002, 0.017, 87.6, 104),
 Wqo = c(0.02, 0.02, 90, 90),
 Sf = c(0.38, 0.38, 70, 70),
 Eightieth = c(0.0154, 0.0154, 101.12, 101.12),
 Twentieth = c(0.0026, 0.0026, 75.84, 75.84)
)

x <- x |> 
  value_to_score(
    value = Value,
    value_type = "Water Quality",
    water_type = "Freshwater",
    indicator = Indicator,
    wqo = Wqo,
    sf = Sf,
    eightieth = Eightieth,
    twentieth = Twentieth
  )
 
}
